use build tags for fargate
